requirements.txt
Flask==3.0.2
spotipy==2.23.0
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
Werkzeug==3.0.1 